+++ apps/gyrokinetic.c
@@ -440,10 +440,11 @@ gkyl_gyrokinetic_app_new(struct gkyl_gk *gk)
   return app;
 }
 
-// Compute fields.
-static void
-calc_field(gkyl_gyrokinetic_app* app, double tcurr, const struct gkyl_array *fin[])
+void
+gyrokinetic_calc_field(gkyl_gyrokinetic_app* app, double tcurr, const struct gkyl_array *fin[])
 {
+  // Compute fields.
+
   if (app->update_field) {
     // Compute electrostatic potential from gyrokinetic Poisson's equation.
     gk_field_accumulate_rho_c(app, app->field, fin);
@@ -466,15 +467,16 @@ calc_field(gkyl_gyrokinetic_app* app, double tcurr, const struct gkyl_array *fin
   }
 }
 
-// Compute fields and apply BCs.
-static void
-calc_field_and_apply_bc(gkyl_gyrokinetic_app* app, double tcurr, struct gkyl_array *distf[], struct gkyl_array *distf_neut[])
+void
+gyrokinetic_calc_field_and_apply_bc(gkyl_gyrokinetic_app* app, double tcurr,
+  struct gkyl_array *distf[], struct gkyl_array *distf_neut[])
 {
+  // Compute fields and apply BCs.
 
   // Compute the field.
   // MF 2024/09/27/: Need the cast here for consistency. Fixing
   // this may require removing 'const' from a lot of places.
-  calc_field(app, tcurr, (const struct gkyl_array **) distf);
+  gyrokinetic_calc_field(app, tcurr, (const struct gkyl_array **) distf);
 
   // Apply boundary conditions.
   for (int i=0; i<app->num_species; ++i) {
@@ -559,7 +561,7 @@ gkyl_gyrokinetic_app_apply_ic(gkyl_gyrokinetic_app* app, double t0)
       gk_species_bflux_rhs(app, s, &s->bflux, distf[i], distf[i]);
     }
   }
-  calc_field_and_apply_bc(app, 0., distf, distf_neut);
+  gyrokinetic_calc_field_and_apply_bc(app, 0., distf, distf_neut);
 }
 
 void
@@ -2122,19 +2124,12 @@ gkyl_gyrokinetic_app_write(gkyl_gyrokinetic_app* app, double tm, int frame)
 // ............. End of write functions ............... //
 // 
 
-// Take a forward Euler step with the suggested time-step dt. This may
-// not be the actual time-step taken. However, the function will never
-// take a time-step larger than dt even if it is allowed by
-// stability. The actual time-step and dt_suggested are returned in
-// the status object.
-static void
-forward_euler(gkyl_gyrokinetic_app* app, double tcurr, double dt,
+void
+gyrokinetic_rhs(gkyl_gyrokinetic_app* app, double tcurr, double dt,
   const struct gkyl_array *fin[], struct gkyl_array *fout[], 
   const struct gkyl_array *fin_neut[], struct gkyl_array *fout_neut[], 
   struct gkyl_update_status *st)
 {
-  app->stat.nfeuler += 1;
-
   double dtmin = DBL_MAX;
 
   // Compute necessary moments and boundary corrections for collisions.
@@ -2240,224 +2235,6 @@ forward_euler(gkyl_gyrokinetic_app* app, double tcurr, double dt,
   double dta = st->dt_actual = dt < dtmin ? dt : dtmin;
   st->dt_suggested = dtmin;
 
-  // Complete update of distribution functions.
-  for (int i=0; i<app->num_species; ++i) {
-    gkyl_array_accumulate(gkyl_array_scale(fout[i], dta), 1.0, fin[i]);
-  }
-  for (int i=0; i<app->num_neut_species; ++i) {
-    if (!app->neut_species[i].info.is_static) {
-      gkyl_array_accumulate(gkyl_array_scale(fout_neut[i], dta), 1.0, fin_neut[i]);
-    }
-  }
-
-}
-
-// Take time-step using the RK3 method. Also sets the status object
-// which has the actual and suggested dts used. These can be different
-// from the actual time-step.
-static struct gkyl_update_status
-rk3(gkyl_gyrokinetic_app* app, double dt0)
-{
-  const struct gkyl_array *fin[app->num_species];
-  struct gkyl_array *fout[app->num_species];
-  const struct gkyl_array *fin_neut[app->num_neut_species];
-  struct gkyl_array *fout_neut[app->num_neut_species];
-  struct gkyl_update_status st = { .success = true };
-
-  // time-stepper state
-  enum { RK_STAGE_1, RK_STAGE_2, RK_STAGE_3, RK_COMPLETE } state = RK_STAGE_1;
-
-  double tcurr = app->tcurr, dt = dt0;
-  while (state != RK_COMPLETE) {
-    switch (state) {
-      case RK_STAGE_1:
-        for (int i=0; i<app->num_species; ++i) {
-          fin[i] = app->species[i].f;
-          fout[i] = app->species[i].f1;
-        }
-        for (int i=0; i<app->num_neut_species; ++i) {
-          fin_neut[i] = app->neut_species[i].f;
-          if (!app->neut_species[i].info.is_static) {
-            fout_neut[i] = app->neut_species[i].f1;
-          }
-        }
-
-        forward_euler(app, tcurr, dt, fin, fout, fin_neut, fout_neut, &st);
-        // Compute the fields and apply BCs.
-        calc_field_and_apply_bc(app, tcurr, fout, fout_neut);
-
-        dt = st.dt_actual;
-        state = RK_STAGE_2;
-        break;
-
-      case RK_STAGE_2:
-        for (int i=0; i<app->num_species; ++i) {
-          fin[i] = app->species[i].f1;
-          fout[i] = app->species[i].fnew;
-        }
-        for (int i=0; i<app->num_neut_species; ++i) {
-          if (!app->neut_species[i].info.is_static) {
-            fin_neut[i] = app->neut_species[i].f1;
-            fout_neut[i] = app->neut_species[i].fnew;
-          }
-          else {
-            fin_neut[i] = app->neut_species[i].f;
-          }
-        }
-
-        forward_euler(app, tcurr+dt, dt, fin, fout, fin_neut, fout_neut, &st);
-
-        if (st.dt_actual < dt) {
-
-          // Recalculate the field.
-          for (int i=0; i<app->num_species; ++i)
-            fin[i] = app->species[i].f;
-          calc_field(app, tcurr, fin);
-
-          // collect stats
-          double dt_rel_diff = (dt-st.dt_actual)/st.dt_actual;
-          app->stat.stage_2_dt_diff[0] = fmin(app->stat.stage_2_dt_diff[0],
-            dt_rel_diff);
-          app->stat.stage_2_dt_diff[1] = fmax(app->stat.stage_2_dt_diff[1],
-            dt_rel_diff);
-          app->stat.nstage_2_fail += 1;
-
-          dt = st.dt_actual;
-          state = RK_STAGE_1; // restart from stage 1
-
-        } 
-        else {
-          for (int i=0; i<app->num_species; ++i) {
-            array_combine(app->species[i].f1,
-              3.0/4.0, app->species[i].f, 1.0/4.0, app->species[i].fnew, &app->species[i].local_ext);
-          }
-          for (int i=0; i<app->num_neut_species; ++i) {
-            if (!app->neut_species[i].info.is_static) {
-              array_combine(app->neut_species[i].f1,
-                3.0/4.0, app->neut_species[i].f, 1.0/4.0, app->neut_species[i].fnew, &app->neut_species[i].local_ext);
-            }
-          }
-
-          // Compute the fields and apply BCs.
-          for (int i=0; i<app->num_species; ++i) {
-            fout[i] = app->species[i].f1;
-          }
-          for (int i=0; i<app->num_neut_species; ++i) {
-            fout_neut[i] = app->neut_species[i].f1;
-          }
-          calc_field_and_apply_bc(app, tcurr, fout, fout_neut);
-
-          state = RK_STAGE_3;
-        }
-        break;
-
-      case RK_STAGE_3:
-        for (int i=0; i<app->num_species; ++i) {
-          fin[i] = app->species[i].f1;
-          fout[i] = app->species[i].fnew;
-        }
-        for (int i=0; i<app->num_neut_species; ++i) {
-          if (!app->neut_species[i].info.is_static) {
-            fin_neut[i] = app->neut_species[i].f1;
-            fout_neut[i] = app->neut_species[i].fnew;
-          }
-          else {
-            fin_neut[i] = app->neut_species[i].f;
-          }          
-        }
-
-        forward_euler(app, tcurr+dt/2, dt, fin, fout, fin_neut, fout_neut, &st);
-
-        if (st.dt_actual < dt) {
-          // Recalculate the field.
-          for (int i=0; i<app->num_species; ++i)
-            fin[i] = app->species[i].f;
-          calc_field(app, tcurr, fin);
-
-          // collect stats
-          double dt_rel_diff = (dt-st.dt_actual)/st.dt_actual;
-          app->stat.stage_3_dt_diff[0] = fmin(app->stat.stage_3_dt_diff[0],
-            dt_rel_diff);
-          app->stat.stage_3_dt_diff[1] = fmax(app->stat.stage_3_dt_diff[1],
-            dt_rel_diff);
-          app->stat.nstage_3_fail += 1;
-
-          dt = st.dt_actual;
-          state = RK_STAGE_1; // restart from stage 1
-
-          app->stat.nstage_2_fail += 1;
-        }
-        else {
-          for (int i=0; i<app->num_species; ++i) {
-            array_combine(app->species[i].f1,
-              1.0/3.0, app->species[i].f, 2.0/3.0, app->species[i].fnew, &app->species[i].local_ext);
-            gkyl_array_copy_range(app->species[i].f, app->species[i].f1, &app->species[i].local_ext);
-          }
-          for (int i=0; i<app->num_neut_species; ++i) {
-            if (!app->neut_species[i].info.is_static) {
-              array_combine(app->neut_species[i].f1,
-                1.0/3.0, app->neut_species[i].f, 2.0/3.0, app->neut_species[i].fnew, &app->neut_species[i].local_ext);
-              gkyl_array_copy_range(app->neut_species[i].f, app->neut_species[i].f1, &app->neut_species[i].local_ext);
-            }
-          }
-
-          if (app->enforce_positivity) {
-            // Apply positivity shift if requested.
-            int elc_idx = -1;
-            gkyl_array_clear(app->ps_delta_m0_ions, 0.0);
-            for (int i=0; i<app->num_species; ++i) {
-              struct gk_species *gks = &app->species[i];
-
-              // Copy f so we can calculate the moments of the change later. 
-              gkyl_array_set(gks->fnew, -1.0, gks->f);
-
-              // Shift each species.
-              gkyl_positivity_shift_gyrokinetic_advance(gks->pos_shift_op, &app->local, &gks->local,
-                gks->f, gks->m0.marr, gks->ps_delta_m0);
-
-              // Accumulate the shift density of all ions:
-              if (gks->info.charge > 0.0)
-                gkyl_array_accumulate(app->ps_delta_m0_ions, 1.0, gks->ps_delta_m0);
-              else if (gks->info.charge < 0.0) 
-                elc_idx = i;
-            }
-
-            // Rescale each species to enforce quasineutrality.
-            for (int i=0; i<app->num_species; ++i) {
-              struct gk_species *gks = &app->species[i];
-              if (gks->info.charge > 0.0) {
-                struct gk_species *gkelc = &app->species[elc_idx];
-                gkyl_positivity_shift_gyrokinetic_quasineutrality_scale(gks->pos_shift_op, &app->local, &gks->local,
-                  gks->ps_delta_m0, app->ps_delta_m0_ions, gkelc->ps_delta_m0, gks->m0.marr, gks->f);
-              }
-              else {
-                gkyl_positivity_shift_gyrokinetic_quasineutrality_scale(gks->pos_shift_op, &app->local, &gks->local,
-                  gks->ps_delta_m0, gks->ps_delta_m0, app->ps_delta_m0_ions, gks->m0.marr, gks->f);
-              }
-
-              gkyl_array_accumulate(gks->fnew, 1.0, gks->f);
-            }
-          }
-
-          // Compute the fields and apply BCs
-          for (int i=0; i<app->num_species; ++i) {
-            fout[i] = app->species[i].f;
-          }
-          for (int i=0; i<app->num_neut_species; ++i) {
-            fout_neut[i] = app->neut_species[i].f;
-          }
-          calc_field_and_apply_bc(app, tcurr, fout, fout_neut);
-
-          state = RK_COMPLETE;
-        }
-        break;
-
-      case RK_COMPLETE: // can't happen: suppresses warning
-        break;
-    }
-  }
-
-  return st;
 }
 
 struct gkyl_update_status
@@ -2466,7 +2243,7 @@ gkyl_gyrokinetic_update(gkyl_gyrokinetic_app* app, double dt)
   app->stat.nup += 1;
   struct timespec wst = gkyl_wall_clock();
 
-  struct gkyl_update_status status = rk3(app, dt);
+  struct gkyl_update_status status = gyrokinetic_update_ssp_rk3(app, dt);
   app->tcurr += status.dt_actual;
 
   app->stat.total_tm += gkyl_time_diff_now_sec(wst);
@@ -2914,7 +2691,7 @@ gkyl_gyrokinetic_app_read_from_frame(gkyl_gyrokinetic_app *app, int frame)
         gk_species_bflux_rhs(app, s, &s->bflux, distf[i], distf[i]);
       }
     }
-    calc_field_and_apply_bc(app, rstat.stime, distf, distf_neut);
+    gyrokinetic_calc_field_and_apply_bc(app, rstat.stime, distf, distf_neut);
   }
 
   app->field->is_first_energy_write_call = false; // Append to existing diagnostic.
